+++ docs/changes.rst
@@ -14,6 +14,7 @@ Not yet released.
 * Highlight placeables in translation editor.
 * Improved support for non standard language codes.
 * Added alert on using ambiguous language codes.
+* User is now presented filtered list of languages when adding new translation.
 
 Weblate 4.3.2
 -------------